@@ -1,3 +1,6 @@
+/**
+ * I18n translation class.
+ */
 class I18n {
   constructor({ locale, translations }) {
     this.locale = locale;
@@ -6,16 +9,19 @@ class I18n {
   }
 
   t(str) {
+    const locale = this.locale || this.fallbackLocale;
+
     if (!this.translations[str]) {
       throw new Error(`${str} Translation string not found`);
     }
 
-    if (!this.translations[str][this.locale || this.fallbackLocale]) {
-      throw new Error(`${str} Translation locale not found`);
+    if (!this.translations[str][locale]) {
+      throw new Error(`'${str}' translation not found for locale '${locale}'`);
     }
 
-    return this.translations[str][this.locale || this.fallbackLocale];
+    return this.translations[str][locale];
   }
 }
 
-module.exports = I18n;
+export { I18n };
+export default I18n;
