@@ -1,3 +1,4 @@
 const blacklist = ["renovate-bot", "technote-space", "sw-yx"];
 
-module.exports = blacklist;
+export { blacklist };
+export default blacklist;
