@@ -1,7 +1,22 @@
-const { logger, CustomError } = require("../common/utils");
+import { CustomError, logger } from "./utils.js";
 
+// Script variables.
+const PATs = Object.keys(process.env).filter((key) =>
+  /PAT_\d*$/.exec(key),
+).length;
+const RETRIES = PATs ? PATs : 7;
+
+/**
+ * Try to execute the fetcher function until it succeeds or the max number of retries is reached.
+ *
+ * @param {object[]} retryerParams Object that contains the createTextNode parameters.
+ * @param {object[]} retryerParams.fetcher The fetcher function.
+ * @param {object[]} retryerParams.variables Object with arguments to pass to the fetcher function.
+ * @param {number} retryerParams.retries How many times to retry.
+ * @returns Promise<retryer>
+ */
 const retryer = async (fetcher, variables, retries = 0) => {
-  if (retries > 7) {
+  if (retries > RETRIES) {
     throw new CustomError("Maximum retries exceeded", CustomError.MAX_RETRY);
   }
   try {
@@ -30,14 +45,20 @@ const retryer = async (fetcher, variables, retries = 0) => {
     // prettier-ignore
     // also checking for bad credentials if any tokens gets invalidated
     const isBadCredential = err.response.data && err.response.data.message === "Bad credentials";
+    const isAccountSuspended =
+      err.response.data &&
+      err.response.data.message === "Sorry. Your account was suspended.";
 
-    if (isBadCredential) {
+    if (isBadCredential || isAccountSuspended) {
       logger.log(`PAT_${retries + 1} Failed`);
       retries++;
       // directly return from the function
       return retryer(fetcher, variables, retries);
+    } else {
+      return err.response;
     }
   }
 };
 
-module.exports = retryer;
+export { retryer };
+export default retryer;
