@@ -1,23 +1,30 @@
 // @ts-check
-const axios = require("axios");
-const wrap = require("word-wrap");
-const themes = require("../../themes");
-const toEmoji = require("emoji-name-map");
+import axios from "axios";
+import toEmoji from "emoji-name-map";
+import wrap from "word-wrap";
+import { themes } from "../../themes/index.js";
+
+// Script parameters.
+const ERROR_CARD_LENGTH = 576.5;
 
 /**
- * @param {string} message
- * @param {string} secondaryMessage
- * @returns {string}
+ * Renders error message on the card.
+ *
+ * @param {string} message Main error message.
+ * @param {string} secondaryMessage The secondary error message.
+ * @returns {string} The SVG markup.
  */
 const renderError = (message, secondaryMessage = "") => {
   return `
-    <svg width="495" height="120" viewBox="0 0 495 120" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <svg width="${ERROR_CARD_LENGTH}" height="120" viewBox="0 0 ${ERROR_CARD_LENGTH} 120" fill="none" xmlns="http://www.w3.org/2000/svg">
     <style>
     .text { font: 600 16px 'Segoe UI', Ubuntu, Sans-Serif; fill: #2F80ED }
     .small { font: 600 12px 'Segoe UI', Ubuntu, Sans-Serif; fill: #252525 }
     .gray { fill: #858585 }
     </style>
-    <rect x="0.5" y="0.5" width="494" height="99%" rx="4.5" fill="#FFFEFE" stroke="#E4E2E2"/>
+    <rect x="0.5" y="0.5" width="${
+      ERROR_CARD_LENGTH - 1
+    }" height="99%" rx="4.5" fill="#FFFEFE" stroke="#E4E2E2"/>
     <text x="25" y="45" class="text">Something went wrong! file an issue at https://tiny.one/readme-stats</text>
     <text data-testid="message" x="25" y="55" class="text small">
       <tspan x="25" dy="18">${encodeHTML(message)}</tspan>
@@ -28,86 +35,110 @@ const renderError = (message, secondaryMessage = "") => {
 };
 
 /**
+ * Encode string as HTML.
+ *
  * @see https://stackoverflow.com/a/48073476/10629172
- * @param {string} str
- * @returns {string}
+ *
+ * @param {string} str String to encode.
+ * @returns {string} Encoded string.
  */
-function encodeHTML(str) {
+const encodeHTML = (str) => {
   return str
     .replace(/[\u00A0-\u9999<>&](?!#)/gim, (i) => {
       return "&#" + i.charCodeAt(0) + ";";
     })
     .replace(/\u0008/gim, "");
-}
+};
 
 /**
- * @param {number} num
+ * Retrieves num with suffix k(thousands) precise to 1 decimal if greater than 999.
+ *
+ * @param {number} num The number to format.
+ * @returns {string|number} The formatted number.
  */
-function kFormatter(num) {
+const kFormatter = (num) => {
   return Math.abs(num) > 999
     ? Math.sign(num) * parseFloat((Math.abs(num) / 1000).toFixed(1)) + "k"
     : Math.sign(num) * Math.abs(num);
-}
+};
 
 /**
- * @param {string} hexColor
- * @returns {boolean}
+ * Checks if a string is a valid hex color.
+ *
+ * @param {string} hexColor String to check.
+ * @returns {boolean} True if the given string is a valid hex color.
  */
-function isValidHexColor(hexColor) {
+const isValidHexColor = (hexColor) => {
   return new RegExp(
     /^([A-Fa-f0-9]{8}|[A-Fa-f0-9]{6}|[A-Fa-f0-9]{3}|[A-Fa-f0-9]{4})$/,
   ).test(hexColor);
-}
+};
 
 /**
- * @param {string} value
- * @returns {boolean | string}
+ * Returns boolean if value is either "true" or "false" else the value as it is.
+ *
+ * @param {string | boolean} value The value to parse.
+ * @returns {boolean | undefined } The parsed value.
  */
-function parseBoolean(value) {
-  if (value === "true") {
-    return true;
-  } else if (value === "false") {
-    return false;
-  } else {
-    return value;
+const parseBoolean = (value) => {
+  if (typeof value === "boolean") return value;
+
+  if (typeof value === "string") {
+    if (value.toLowerCase() === "true") {
+      return true;
+    } else if (value.toLowerCase() === "false") {
+      return false;
+    }
   }
-}
+  return undefined;
+};
 
 /**
- * @param {string} str
+ * Parse string to array of strings.
+ *
+ * @param {string} str The string to parse.
+ * @returns {string[]} The array of strings.
  */
-function parseArray(str) {
+const parseArray = (str) => {
   if (!str) return [];
   return str.split(",");
-}
+};
 
 /**
- * @param {number} number
- * @param {number} min
- * @param {number} max
+ * Clamp the given number between the given range.
+ *
+ * @param {number} number The number to clamp.
+ * @param {number} min The minimum value.
+ * @param {number} max The maximum value.
+ * @returns {number} The clamped number.
  */
-function clampValue(number, min, max) {
+const clampValue = (number, min, max) => {
   // @ts-ignore
   if (Number.isNaN(parseInt(number))) return min;
   return Math.max(min, Math.min(number, max));
-}
+};
 
 /**
- * @param {string[]} colors
+ * Check if the given string is a valid gradient.
+ *
+ * @param {string[]} colors Array of colors.
+ * @returns {boolean} True if the given string is a valid gradient.
  */
-function isValidGradient(colors) {
+const isValidGradient = (colors) => {
   return isValidHexColor(colors[1]) && isValidHexColor(colors[2]);
-}
+};
 
 /**
- * @param {string} color
- * @param {string} fallbackColor
- * @returns {string | string[]}
+ * Retrieves a gradient if color has more than one valid hex codes else a single color.
+ *
+ * @param {string} color The color to parse.
+ * @param {string} fallbackColor The fallback color.
+ * @returns {string | string[]} The gradient or color.
  */
-function fallbackColor(color, fallbackColor) {
-  let colors = color.split(",");
+const fallbackColor = (color, fallbackColor) => {
   let gradient = null;
 
+  let colors = color ? color.split(",") : [];
   if (colors.length > 1 && isValidGradient(colors)) {
     gradient = colors;
   }
@@ -116,13 +147,16 @@ function fallbackColor(color, fallbackColor) {
     (gradient ? gradient : isValidHexColor(color) && `#${color}`) ||
     fallbackColor
   );
-}
+};
 
 /**
- * @param {import('axios').AxiosRequestConfig['data']} data
- * @param {import('axios').AxiosRequestConfig['headers']} headers
+ * Send GraphQL request to GitHub API.
+ *
+ * @param {import('axios').AxiosRequestConfig['data']} data Request data.
+ * @param {import('axios').AxiosRequestConfig['headers']} headers Request headers.
+ * @returns {Promise<any>} Request response.
  */
-function request(data, headers) {
+const request = (data, headers) => {
   // @ts-ignore
   return axios({
     url: "https://api.github.com/graphql",
@@ -130,22 +164,20 @@ function request(data, headers) {
     headers,
     data,
   });
-}
+};
 
 /**
- * @param {object} props
- * @param {string[]} props.items
- * @param {number} props.gap
- * @param {number[]?=} props.sizes
- * @param {"column" | "row"?=} props.direction
+ * Auto layout utility, allows us to layout things vertically or horizontally with
+ * proper gaping.
  *
- * @returns {string[]}
- *
- * @description
- * Auto layout utility, allows us to layout things
- * vertically or horizontally with proper gaping
+ * @param {object} props Function properties.
+ * @param {string[]} props.items Array of items to layout.
+ * @param {number} props.gap Gap between items.
+ * @param {number[]?=} props.sizes Array of sizes for each item.
+ * @param {"column" | "row"?=} props.direction Direction to layout items.
+ * @returns {string[]} Array of items with proper layout.
  */
-function flexLayout({ items, gap, direction, sizes = [] }) {
+const flexLayout = ({ items, gap, direction, sizes = [] }) => {
   let lastSize = 0;
   // filter() for filtering out empty strings
   return items.filter(Boolean).map((item, i) => {
@@ -157,31 +189,31 @@ function flexLayout({ items, gap, direction, sizes = [] }) {
     lastSize += size + gap;
     return `<g transform="${transform}">${item}</g>`;
   });
-}
+};
 
 /**
- * @typedef {object} CardColors
- * @prop {string?=} title_color
- * @prop {string?=} text_color
- * @prop {string?=} icon_color
- * @prop {string?=} bg_color
- * @prop {string?=} border_color
- * @prop {keyof typeof import('../../themes')?=} fallbackTheme
- * @prop {keyof typeof import('../../themes')?=} theme
+ * Returns theme based colors with proper overrides and defaults.
+ *
+ * @param {Object} args Function arguments.
+ * @param {string} args.title_color Card title color.
+ * @param {string} args.text_color Card text color.
+ * @param {string} args.icon_color Card icon color.
+ * @param {string} args.bg_color Card background color.
+ * @param {string} args.border_color Card border color.
+ * @param {string} args.ring_color Card ring color.
+ * @param {string} args.theme Card theme.
+ * @param {string} args.fallbackTheme Fallback theme.
  */
-/**
- * returns theme based colors with proper overrides and defaults
- * @param {CardColors} options
- */
-function getCardColors({
+const getCardColors = ({
   title_color,
   text_color,
   icon_color,
   bg_color,
   border_color,
+  ring_color,
   theme,
   fallbackTheme = "default",
-}) {
+}) => {
   const defaultTheme = themes[fallbackTheme];
   const selectedTheme = themes[theme] || defaultTheme;
   const defaultBorderColor =
@@ -192,6 +224,13 @@ function getCardColors({
   const titleColor = fallbackColor(
     title_color || selectedTheme.title_color,
     "#" + defaultTheme.title_color,
+  );
+
+  // get the color provided by the user else the theme color
+  // finally if both colors are invalid we use the titleColor
+  const ringColor = fallbackColor(
+    ring_color || selectedTheme.ring_color,
+    titleColor,
   );
   const iconColor = fallbackColor(
     icon_color || selectedTheme.icon_color,
@@ -211,16 +250,18 @@ function getCardColors({
     "#" + defaultBorderColor,
   );
 
-  return { titleColor, iconColor, textColor, bgColor, borderColor };
-}
+  return { titleColor, iconColor, textColor, bgColor, borderColor, ringColor };
+};
 
 /**
- * @param {string} text
- * @param {number} width
- * @param {number} maxLines
- * @returns {string[]}
+ * Split text over multiple lines based on the card width.
+ *
+ * @param {string} text Text to split.
+ * @param {number} width Line width in number of characters.
+ * @param {number} maxLines Maximum number of lines.
+ * @returns {string[]} Array of lines.
  */
-function wrapTextMultiline(text, width = 59, maxLines = 3) {
+const wrapTextMultiline = (text, width = 59, maxLines = 3) => {
   const fullWidthComma = "，";
   const encoded = encodeHTML(text);
   const isChinese = encoded.includes(fullWidthComma);
@@ -245,7 +286,7 @@ function wrapTextMultiline(text, width = 59, maxLines = 3) {
   // Remove empty lines if text fits in less than maxLines lines
   const multiLineText = lines.filter(Boolean);
   return multiLineText;
-}
+};
 
 const noop = () => {};
 // return console instance based on the environment
@@ -263,12 +304,17 @@ const SECONDARY_ERROR_MESSAGES = {
   MAX_RETRY:
     "Please add an env variable called PAT_1 with your github token in vercel",
   USER_NOT_FOUND: "Make sure the provided username is not an organization",
+  GRAPHQL_ERROR: "Please try again later",
+  WAKATIME_USER_NOT_FOUND: "Make sure you have a public WakaTime profile",
 };
 
+/**
+ * Custom error class to handle custom GRS errors.
+ */
 class CustomError extends Error {
   /**
-   * @param {string} message
-   * @param {string} type
+   * @param {string} message Error message.
+   * @param {string} type Error type.
    */
   constructor(message, type) {
     super(message);
@@ -278,8 +324,13 @@ class CustomError extends Error {
 
   static MAX_RETRY = "MAX_RETRY";
   static USER_NOT_FOUND = "USER_NOT_FOUND";
+  static GRAPHQL_ERROR = "GRAPHQL_ERROR";
+  static WAKATIME_ERROR = "WAKATIME_ERROR";
 }
 
+/**
+ * Missing query parameter class.
+ */
 class MissingParamError extends Error {
   /**
    * @param {string[]} missedParams
@@ -296,12 +347,14 @@ class MissingParamError extends Error {
 }
 
 /**
+ * Retrieve text length.
+ *
  * @see https://stackoverflow.com/a/48172630/10629172
- * @param {string} str
- * @param {number} fontSize
- * @returns
+ * @param {string} str String to measure.
+ * @param {number} fontSize Font size.
+ * @returns {number} Text length.
  */
-function measureText(str, fontSize = 10) {
+const measureText = (str, fontSize = 10) => {
   // prettier-ignore
   const widths = [
     0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
@@ -333,18 +386,20 @@ function measureText(str, fontSize = 10) {
       )
       .reduce((cur, acc) => acc + cur) * fontSize
   );
-}
+};
 
 /** @param {string} name */
 const lowercaseTrim = (name) => name.toLowerCase().trim();
 
 /**
- * @template T
- * @param {Array<T>} arr
- * @param {number} perChunk
- * @returns {Array<T>}
+ * Split array of languages in two columns.
+ *
+ * @template T Language object.
+ * @param {Array<T>} arr Array of languages.
+ * @param {number} perChunk Number of languages per column.
+ * @returns {Array<T>} Array of languages split in two columns.
  */
-function chunkArray(arr, perChunk) {
+const chunkArray = (arr, perChunk) => {
   return arr.reduce((resultArray, item, index) => {
     const chunkIndex = Math.floor(index / perChunk);
 
@@ -356,39 +411,56 @@ function chunkArray(arr, perChunk) {
 
     return resultArray;
   }, []);
-}
+};
 
 /**
+ * Parse emoji from string.
  *
- * @param {string} str
- * @returns {string}
+ * @param {string} str String to parse emoji from.
+ * @returns {string} String with emoji parsed.
  */
-function parseEmojis(str) {
+const parseEmojis = (str) => {
   if (!str) throw new Error("[parseEmoji]: str argument not provided");
   return str.replace(/:\w+:/gm, (emoji) => {
     return toEmoji.get(emoji) || "";
   });
-}
+};
 
-module.exports = {
+/**
+ * Get diff in minutes
+ * @param {Date} d1
+ * @param {Date} d2
+ * @returns {number}
+ */
+const dateDiff = (d1, d2) => {
+  const date1 = new Date(d1);
+  const date2 = new Date(d2);
+  const diff = date1.getTime() - date2.getTime();
+  return Math.round(diff / (1000 * 60));
+};
+
+export {
+  ERROR_CARD_LENGTH,
   renderError,
-  kFormatter,
   encodeHTML,
+  kFormatter,
   isValidHexColor,
-  request,
-  parseArray,
   parseBoolean,
+  parseArray,
+  clampValue,
+  isValidGradient,
   fallbackColor,
+  request,
   flexLayout,
   getCardColors,
-  clampValue,
   wrapTextMultiline,
-  measureText,
   logger,
   CONSTANTS,
   CustomError,
   MissingParamError,
+  measureText,
   lowercaseTrim,
   chunkArray,
   parseEmojis,
+  dateDiff,
 };
