@@ -1,10 +1,13 @@
 // @ts-check
-const retryer = require("../common/retryer");
-const { request, MissingParamError } = require("../common/utils");
+import { retryer } from "../common/retryer.js";
+import { MissingParamError, request } from "../common/utils.js";
 
 /**
- * @param {import('Axios').AxiosRequestHeaders} variables
- * @param {string} token
+ * Repo data fetcher.
+ *
+ * @param {import('axios').AxiosRequestHeaders} variables Fetcher variables.
+ * @param {string} token GitHub token.
+ * @returns {Promise<import('axios').AxiosResponse>} The response.
  */
 const fetcher = (variables, token) => {
   return request(
@@ -51,11 +54,13 @@ const fetcher = (variables, token) => {
 const urlExample = "/api/pin?username=USERNAME&amp;repo=REPO_NAME";
 
 /**
- * @param {string} username
- * @param {string} reponame
- * @returns {Promise<import("./types").RepositoryData>}
+ * Fetch repository data.
+ *
+ * @param {string} username GitHub username.
+ * @param {string} reponame GitHub repository name.
+ * @returns {Promise<import("./types").RepositoryData>} Repository data.
  */
-async function fetchRepo(username, reponame) {
+const fetchRepo = async (username, reponame) => {
   if (!username && !reponame) {
     throw new MissingParamError(["username", "repo"], urlExample);
   }
@@ -95,6 +100,7 @@ async function fetchRepo(username, reponame) {
       starCount: data.organization.repository.stargazers.totalCount,
     };
   }
-}
+};
 
-module.exports = fetchRepo;
+export { fetchRepo };
+export default fetchRepo;
