@@ -1,5 +1,14 @@
-const { encodeHTML } = require("./common/utils");
+import { encodeHTML } from "./common/utils.js";
 
+/**
+ * Retrieves stat card labels in the available locales.
+ *
+ * @param {string} name The name of the locale.
+ * @param {string} apostrophe Whether to use apostrophe or not.
+ * @returns {Object} The locales object.
+ *
+ * @see https://www.andiamo.co.uk/resources/iso-language-codes/ for language codes.
+ */
 const statCardLocales = ({ name, apostrophe }) => {
   const encodedName = encodeHTML(name);
   return {
@@ -30,7 +39,7 @@ const statCardLocales = ({ name, apostrophe }) => {
       sk: `GitHub štatistiky používateľa ${encodedName}`,
       tr: `${encodedName} Hesabının GitHub Yıldızları`,
       pl: `Statystyki GitHub użytkownika ${encodedName}`,
-      uz: `${encodedName}ning Github'dagi statistikasi`,
+      uz: `${encodedName}ning GitHub'dagi statistikasi`,
       vi: `Thống Kê GitHub ${encodedName}`,
       se: `GitHubstatistik för ${encodedName}`,
     },
@@ -60,7 +69,7 @@ const statCardLocales = ({ name, apostrophe }) => {
       my: "Jumlah Bintang",
       sk: "Hviezdy",
       tr: "Toplam Yıldız",
-      pl: "Liczba Gwiazdek dostanych",
+      pl: "Liczba otrzymanych gwiazdek",
       uz: "Yulduzchalar",
       vi: "Tổng Số Sao",
       se: "Antal intjänade stjärnor",
@@ -153,41 +162,131 @@ const statCardLocales = ({ name, apostrophe }) => {
       my: "Jumlah Isu Dilaporkan",
       sk: "Všetky problémy",
       tr: "Toplam Hata",
-      pl: "Wszystkie Issues",
+      pl: "Wszystkie problemy",
       uz: "'Issue'lar",
       vi: "Tổng Số Vấn Đề",
       se: "Total antal issues",
     },
     "statcard.contribs": {
-      ar: "ساهم في",
-      cn: "参与项目数",
-      "zh-tw": "參與項目數",
-      cs: "Přispěl k",
-      de: "Beigetragen zu",
-      en: "Contributed to",
-      bn: "অবদান রেখেছেন",
-      es: "Contribuciones en",
-      fr: "Contribué à",
-      hu: "Hozzájárulások",
-      it: "Ha contribuito a",
-      ja: "貢献したリポジトリ",
-      kr: "전체 기여도",
-      nl: "Bijgedragen aan",
-      "pt-pt": "Contribuiu em",
-      "pt-br": "Contribuiu para",
-      np: "कुल योगदानहरू",
-      el: "Συνεισφέρθηκε σε",
-      ru: "Внёс вклад в",
-      "uk-ua": "Вніс внесок у",
-      id: "Berkontribusi ke",
-      ml: "സമർപ്പിച്ചിരിക്കുന്നത്",
-      my: "Menyumbang kepada",
-      sk: "Účasti",
-      tr: "Katkı Verildi",
-      pl: "Kontrybucje",
-      uz: "Hissa qoʻshgan",
-      vi: "Đã Đóng Góp",
-      se: "Bidragit till",
+      ar: "ساهم في (العام الماضي)",
+      cn: "贡献于（去年）",
+      "zh-tw": "參與項目數 （去年）",
+      cs: "Přispěl k (minulý rok)",
+      de: "Beigetragen zu (letztes Jahr)",
+      en: "Contributed to (last year)",
+      bn: "অবদান (গত বছর)",
+      es: "Contribuciones en (el año pasado)",
+      fr: "Contribué à (l'année dernière)",
+      hu: "Hozzájárulások (tavaly)",
+      it: "Ha contribuito a (l'anno scorso)",
+      ja: "貢献したリポジトリ （昨年）",
+      kr: "(작년) 기여",
+      nl: "Bijgedragen aan (vorig jaar)",
+      "pt-pt": "Contribuiu em (ano passado)",
+      "pt-br": "Contribuiu para (ano passado)",
+      np: "कुल योगदानहरू (गत वर्ष)",
+      el: "Συνεισφέρθηκε σε (πέρυσι)",
+      ru: "Внёс вклад в (за прошлый год)",
+      "uk-ua": "Зробив внесок у (за минулий рік)",
+      id: "Berkontribusi ke (tahun lalu)",
+      ml: "സമർപ്പിച്ചിരിക്കുന്നത് (കഴിഞ്ഞ വർഷം)",
+      my: "Menyumbang kepada (tahun lepas)",
+      sk: "Účasti (minulý rok)",
+      tr: "Katkı Verildi (geçen yıl)",
+      pl: "Kontrybucje (w zeszłym roku)",
+      uz: "Hissa qoʻshgan (o'tgan yili)",
+      vi: "Đã Đóng Góp (năm ngoái)",
+      se: "Bidragit till (förra året)",
+    },
+    "statcard.reviews": {
+      ar: "تمت مراجعة إجمالي العلاقات العامة",
+      cn: "審查的 PR 總數",
+      "zh-tw": "审查的 PR 总数",
+      cs: "Celkový počet PR",
+      de: "Insgesamt überprüfte PRs",
+      en: "Total PRs reviewed",
+      bn: "সর্বমোট পুনরালোচনা করা PR",
+      es: "PR totales revisados",
+      fr: "Nombre total de PR examinés",
+      hu: "Összes ellenőrzött PR",
+      it: "PR totali esaminati",
+      ja: "レビューされた PR の総数",
+      kr: "검토된 총 PR",
+      nl: "Totaal beoordeelde PR's",
+      "pt-pt": "Total de PRs revistos",
+      "pt-br": "Total de PRs revisados",
+      np: "कुल पीआर समीक्षित",
+      el: "Σύνολο Αναθεωρημένων PR",
+      ru: "Всего pull request`ов проверено",
+      "uk-ua": "Всього pull request`iв перевірено",
+      id: "Total PR yang Direview",
+      my: "Jumlah PR Dikaji Semula",
+      sk: "Celkový počet PR",
+      tr: "İncelenen toplam PR",
+      pl: "Łącznie sprawdzonych PR",
+      uz: "Koʻrib chiqilgan PR-lar soni",
+      vi: "Tổng Số PR Đã Xem Xét",
+      se: "Totalt antal granskade PR",
+    },
+    "statcard.discussions-started": {
+      ar: "مجموع بدء المناقشات",
+      cn: "发起的讨论总数",
+      "zh-tw": "發起的討論總數",
+      cs: "Celkem zahájených diskusí",
+      de: "Gesamt gestartete Diskussionen",
+      en: "Total Discussions Started",
+      bn: "সর্বমোট আলোচনা শুরু",
+      es: "Discusiones totales iniciadas",
+      fr: "Nombre total de discussions lancées",
+      hu: "Összes megkezdett megbeszélés",
+      it: "Discussioni totali avviate",
+      ja: "開始されたディスカッションの総数",
+      kr: "시작된 토론 총 수",
+      nl: "Totaal gestarte discussies",
+      "pt-pt": "Total de Discussões Iniciadas",
+      "pt-br": "Total de Discussões Iniciadas",
+      np: "कुल चर्चा सुरु",
+      el: "Σύνολο Συζητήσεων που Ξεκίνησαν",
+      ru: "Всего начатых дискуссий",
+      "uk-ua": "Всього розпочатих дискусій",
+      id: "Total Diskusi Dimulai",
+      my: "Jumlah Perbincangan Bermula",
+      sk: "Celkový počet začatých diskusií",
+      tr: "Başlatılan Toplam Tartışma",
+      pl: "Łącznie rozpoczętych dyskusji",
+      uz: "Boshlangan muzokaralar soni",
+      vi: "Tổng Số Thảo Luận Bắt Đầu",
+      se: "Totalt antal diskussioner startade",
+    },
+    "statcard.discussions-answered": {
+      ar: "مجموع الردود على المناقشات",
+      cn: "回复的讨论总数",
+      "zh-tw": "回覆的討論總數",
+      cs: "Celkem zodpovězených diskusí",
+      de: "Gesamt beantwortete Diskussionen",
+      en: "Total Discussions Answered",
+      bn: "সর্বমোট আলোচনা উত্তর",
+      es: "Discusiones totales respondidas",
+      fr: "Nombre total de discussions répondues",
+      hu: "Összes megválaszolt megbeszélés",
+      it: "Discussioni totali risposte",
+      ja: "回答されたディスカッションの総数",
+      kr: "답변된 토론 총 수",
+      nl: "Totaal beantwoorde discussies",
+      "pt-pt": "Total de Discussões Respondidas",
+      "pt-br": "Total de Discussões Respondidas",
+      np: "कुल चर्चा उत्तर",
+      el: "Σύνολο Συζητήσεων που Απαντήθηκαν",
+      ru: "Всего отвеченных дискуссий",
+      "uk-ua": "Всього відповідей на дискусії",
+      id: "Total Diskusi Dibalas",
+      my: "Jumlah Perbincangan Dijawab",
+      sk: "Celkový počet zodpovedaných diskusií",
+      tr: "Toplam Cevaplanan Tartışma",
+      pl: "Łącznie odpowiedzianych dyskusji",
+      uz: "Javob berilgan muzokaralar soni",
+      vi: "Tổng Số Thảo Luận Đã Trả Lời",
+      se: "Totalt antal diskussioner besvarade",
     },
   };
 };
@@ -276,7 +375,7 @@ const langCardLocales = {
     np: "अधिक प्रयोग गरिएको भाषाहरू",
     el: "Οι περισσότερο χρησιμοποιούμενες γλώσσες",
     ru: "Наиболее часто используемые языки",
-    "uk-ua": "Найбільш часто використовувані мови",
+    "uk-ua": "Найчастіше використовувані мови",
     id: "Bahasa Yang Paling Banyak Digunakan",
     ml: "കൂടുതൽ ഉപയോഗിച്ച ഭാഷകൾ",
     my: "Bahasa Paling Digunakan",
@@ -285,6 +384,36 @@ const langCardLocales = {
     pl: "Najczęściej używane języki",
     vi: "Ngôn Ngữ Thường Sử Dụng",
     se: "Mest använda språken",
+  },
+  "langcard.nodata": {
+    ar: "لا توجد بيانات لغات.",
+    cn: "沒有語言數據。",
+    "zh-tw": "沒有語言數據。",
+    cs: "Žádné jazykové údaje.",
+    de: "Keine Sprachdaten.",
+    bn: "কোন ভাষার ডেটা নেই।",
+    en: "No languages data.",
+    es: "Sin datos de idiomas.",
+    fr: "Aucune donnée sur les langues.",
+    hu: "Nincsenek nyelvi adatok.",
+    it: "Nessun dato sulle lingue.",
+    ja: "言語データがありません。",
+    kr: "언어 데이터가 없습니다.",
+    nl: "Ingen sprogdata.",
+    "pt-pt": "Sem dados de idiomas.",
+    "pt-br": "Sem dados de idiomas.",
+    np: "कुनै भाषा डाटा छैन।",
+    el: "Δεν υπάρχουν δεδομένα γλωσσών.",
+    ru: "Нет данных о языках.",
+    "uk-ua": "Немає даних про мови.",
+    id: "Tidak ada data bahasa.",
+    ml: "ഭാഷാ ഡാറ്റയില്ല.",
+    my: "Tiada data bahasa.",
+    sk: "Žiadne údaje o jazykoch.",
+    tr: "Dil verisi yok.",
+    pl: "Brak danych dotyczących języków.",
+    vi: "Không có dữ liệu ngôn ngữ.",
+    se: "Inga språkdata.",
   },
 };
 
@@ -315,9 +444,131 @@ const wakatimeCardLocales = {
     my: "Statistik Wakatime",
     sk: "Wakatime štatistika",
     tr: "Waketime İstatistikler",
-    pl: "statystyki Wakatime",
+    pl: "Statystyki Wakatime",
     vi: "Thống Kê Wakatime",
     se: "Wakatime statistik",
+  },
+  "wakatimecard.lastyear": {
+    ar: "العام الماضي",
+    cn: "去年",
+    "zh-tw": "去年",
+    cs: "Minulý rok",
+    de: "Letztes Jahr",
+    en: "last year",
+    bn: "গত বছর",
+    es: "El año pasado",
+    fr: "L'année dernière",
+    hu: "Tavaly",
+    it: "L'anno scorso",
+    ja: "昨年",
+    kr: "작년",
+    nl: "Vorig jaar",
+    "pt-pt": "Ano passado",
+    "pt-br": "Ano passado",
+    np: "गत वर्ष",
+    el: "Πέρυσι",
+    ru: "За прошлый год",
+    "uk-ua": "За минулий рік",
+    id: "Tahun lalu",
+    ml: "കഴിഞ്ഞ വർഷം",
+    my: "Tahun lepas",
+    sk: "Minulý rok",
+    tr: "Geçen yıl",
+    pl: "W zeszłym roku",
+    vi: "Năm ngoái",
+    se: "Förra året",
+  },
+  "wakatimecard.last7days": {
+    ar: "آخر 7 أيام",
+    cn: "最近 7 天",
+    "zh-tw": "最近 7 天",
+    cs: "Posledních 7 dní",
+    de: "Letzte 7 Tage",
+    en: "last 7 days",
+    bn: "গত ৭ দিন",
+    es: "Últimos 7 días",
+    fr: "7 derniers jours",
+    hu: "Elmúlt 7 nap",
+    it: "Ultimi 7 giorni",
+    ja: "過去 7 日間",
+    kr: "지난 7 일",
+    nl: "Afgelopen 7 dagen",
+    "pt-pt": "Últimos 7 dias",
+    "pt-br": "Últimos 7 dias",
+    np: "गत ७ दिन",
+    el: "Τελευταίες 7 ημέρες",
+    ru: "Последние 7 дней",
+    "uk-ua": "Останні 7 днів",
+    id: "7 hari terakhir",
+    ml: "കഴിഞ്ഞ 7 ദിവസം",
+    my: "7 hari lepas",
+    sk: "Posledných 7 dní",
+    tr: "Son 7 gün",
+    pl: "Ostatnie 7 dni",
+    vi: "7 ngày qua",
+    se: "Senaste 7 dagarna",
+  },
+  "wakatimecard.notpublic": {
+    ar: "ملف المستخدم غير عام",
+    cn: "Wakatime 用户个人资料未公开",
+    "zh-tw": "Wakatime 使用者個人資料未公開",
+    cs: "Profil uživatele Wakatime není veřejný",
+    de: "Wakatime-Benutzerprofil nicht öffentlich",
+    en: "Wakatime user profile not public",
+    bn: "Wakatime ব্যবহারকারীর প্রোফাইল প্রকাশ্য নয়",
+    es: "Perfil de usuario de Wakatime no público",
+    fr: "Profil utilisateur Wakatime non public",
+    hu: "A Wakatime felhasználói profilja nem nyilvános",
+    it: "Profilo utente Wakatime non pubblico",
+    ja: "Wakatime ユーザープロファイルは公開されていません",
+    kr: "Wakatime 사용자 프로필이 공개되지 않았습니다",
+    nl: "Wakatime gebruikersprofiel niet openbaar",
+    "pt-pt": "Perfil de usuário Wakatime não público",
+    "pt-br": "Perfil de usuário Wakatime não público",
+    np: "Wakatime प्रयोगकर्ता प्रोफाइल सार्वजनिक छैन",
+    el: "Το προφίλ χρήστη Wakatime δεν είναι δημόσιο",
+    ru: "Профиль пользователя Wakatime не является общедоступным",
+    "uk-ua": "Профіль користувача Wakatime не є публічним",
+    id: "Profil pengguna Wakatime tidak publik",
+    ml: "Wakatime ഉപയോക്തൃ പ്രൊഫൈൽ പൊതുവായി പ്രസിദ്ധീകരിക്കപ്പെടാത്തതാണ്",
+    my: "Profil pengguna Wakatime tidak awam",
+    sk: "Profil používateľa Wakatime nie je verejný",
+    tr: "Wakatime kullanıcı profili herkese açık değil",
+    pl: "Profil użytkownika Wakatime nie jest publiczny",
+    vi: "Hồ sơ người dùng Wakatime không công khai",
+    se: "Wakatime användarprofil inte offentlig",
+  },
+  "wakatimecard.nocodedetails": {
+    ar: "المستخدم لا يشارك معلومات تفصيلية عن البرمجة",
+    cn: "用户不公开分享详细的代码统计信息",
+    "zh-tw": "使用者不公開分享詳細的程式碼統計資訊",
+    cs: "Uživatel nesdílí podrobné statistiky kódu",
+    de: "Benutzer teilt keine detaillierten Code-Statistiken",
+    en: "User doesn't publicly share detailed code statistics",
+    bn: "ব্যবহারকারী বিস্তারিত কোড পরিসংখ্যান প্রকাশ করেন না",
+    es: "El usuario no comparte públicamente estadísticas detalladas de código",
+    fr: "L'utilisateur ne partage pas publiquement de statistiques de code détaillées",
+    hu: "A felhasználó nem osztja meg nyilvánosan a részletes kódstatisztikákat",
+    it: "L'utente non condivide pubblicamente statistiche dettagliate sul codice",
+    ja: "ユーザーは詳細なコード統計を公開しません",
+    kr: "사용자는 자세한 코드 통계를 공개하지 않습니다",
+    nl: "Gebruiker deelt geen gedetailleerde code-statistieken",
+    "pt-pt":
+      "O utilizador não partilha publicamente estatísticas detalhadas de código",
+    "pt-br":
+      "O usuário não compartilha publicamente estatísticas detalhadas de código",
+    np: "प्रयोगकर्ता सार्वजनिक रूपमा विस्तृत कोड तथ्याङ्क साझा गर्दैन",
+    el: "Ο χρήστης δεν δημοσιεύει δημόσια λεπτομερείς στατιστικές κώδικα",
+    ru: "Пользователь не делится подробной статистикой кода",
+    "uk-ua": "Користувач не публікує детальну статистику коду",
+    id: "Pengguna tidak membagikan statistik kode terperinci secara publik",
+    ml: "ഉപയോക്താവ് പൊതുവെ വിശദീകരിച്ച കോഡ് സ്റ്റാറ്റിസ്റ്റിക്സ് പങ്കിടുന്നില്ല",
+    my: "Pengguna tidak berkongsi statistik kod terperinci secara awam",
+    sk: "Používateľ neposkytuje verejne podrobné štatistiky kódu",
+    tr: "Kullanıcı ayrıntılı kod istatistiklerini herkese açık olarak paylaşmıyor",
+    pl: "Użytkownik nie udostępnia publicznie szczegółowych statystyk kodu",
+    vi: "Người dùng không chia sẻ thống kê mã chi tiết công khai",
+    se: "Användaren delar inte offentligt detaljerad kodstatistik",
   },
   "wakatimecard.nocodingactivity": {
     ar: "لا يوجد نشاط برمجي لهذا الأسبوع",
@@ -354,15 +605,21 @@ const wakatimeCardLocales = {
 
 const availableLocales = Object.keys(repoCardLocales["repocard.archived"]);
 
-function isLocaleAvailable(locale) {
+/**
+ * Checks whether the locale is available or not.
+ *
+ * @param {string} locale The locale to check.
+ * @returns {boolean} Boolean specifying whether the locale is available or not.
+ */
+const isLocaleAvailable = (locale) => {
   return availableLocales.includes(locale.toLowerCase());
-}
+};
 
-module.exports = {
-  isLocaleAvailable,
+export {
   availableLocales,
-  statCardLocales,
-  repoCardLocales,
+  isLocaleAvailable,
   langCardLocales,
+  repoCardLocales,
+  statCardLocales,
   wakatimeCardLocales,
 };
