@@ -18,11 +18,14 @@ export type RepositoryData = {
 export type StatsData = {
   name: string;
   totalPRs: number;
+  totalReviews: number;
   totalCommits: number;
   totalIssues: number;
   totalStars: number;
+  totalDiscussionsStarted: number;
+  totalDiscussionsAnswered: number;
   contributedTo: number;
-  rank: { level: string; score: number };
+  rank: { level: string; percentile: number };
 };
 
 export type Lang = {
