@@ -1,6 +1,9 @@
 // @ts-check
 /**
- * @param {number} value
+ * Calculates progress along the boundary of the circle i.e it's circumference.
+ *
+ * @param {number} value The rank value to calculate progress for.
+ * @returns {number} Progress value.
  */
 const calculateCircleProgress = (value) => {
   const radius = 40;
@@ -13,9 +16,11 @@ const calculateCircleProgress = (value) => {
 };
 
 /**
+ * Retrieves the animation to display progress along the circumference of circle
+ * from the beginning to the given value in a clockwise direction.
  *
- * @param {{progress: number}} param0
- * @returns
+ * @param {{progress: number}} progress The progress value to animate to.
+ * @returns {string} Progress animation css.
  */
 const getProgressAnimation = ({ progress }) => {
   return `
@@ -30,6 +35,11 @@ const getProgressAnimation = ({ progress }) => {
   `;
 };
 
+/**
+ * Retrieves css animations for a card.
+ *
+ * @returns {string} Animation css.
+ */
 const getAnimations = () => {
   return `
     /* Animations */
@@ -53,18 +63,22 @@ const getAnimations = () => {
 };
 
 /**
- * @param {{
- *  titleColor?: string | string[]
- *  textColor?: string | string[]
- *  iconColor?: string | string[]
- *  show_icons?: boolean;
- *  progress?: number;
- * }} args
+ * Retrieves CSS styles for a card.
+ *
+ * @param {Object} colors The colors to use for the card.
+ * @param {string} colors.titleColor The title color.
+ * @param {string} colors.textColor The text color.
+ * @param {string} colors.iconColor The icon color.
+ * @param {string} colors.ringColor The ring color.
+ * @param {boolean} colors.show_icons Whether to show icons.
+ * @param {number} colors.progress The progress value to animate to.
+ * @returns {string} Card CSS styles.
  */
 const getStyles = ({
   titleColor,
   textColor,
   iconColor,
+  ringColor,
   show_icons,
   progress,
 }) => {
@@ -81,24 +95,31 @@ const getStyles = ({
       animation: fadeInAnimation 0.3s ease-in-out forwards;
     }
     .rank-text {
-      font: 800 24px 'Segoe UI', Ubuntu, Sans-Serif; fill: ${textColor}; 
+      font: 800 24px 'Segoe UI', Ubuntu, Sans-Serif; fill: ${textColor};
       animation: scaleInAnimation 0.3s ease-in-out forwards;
     }
+    .rank-percentile-header {
+      font-size: 14px;
+    }
+    .rank-percentile-text {
+      font-size: 16px;
+    }
     
+    .not_bold { font-weight: 400 }
     .bold { font-weight: 700 }
     .icon {
       fill: ${iconColor};
       display: ${!!show_icons ? "block" : "none"};
     }
-    
+
     .rank-circle-rim {
-      stroke: ${titleColor};
+      stroke: ${ringColor};
       fill: none;
       stroke-width: 6;
       opacity: 0.2;
     }
     .rank-circle {
-      stroke: ${titleColor};
+      stroke: ${ringColor};
       stroke-dasharray: 250;
       fill: none;
       stroke-width: 6;
@@ -112,4 +133,4 @@ const getStyles = ({
   `;
 };
 
-module.exports = { getStyles, getAnimations };
+export { getAnimations, getStyles };
