@@ -1,14 +1,13 @@
-require("@testing-library/jest-dom");
-const {
-  kFormatter,
+import { queryByTestId } from "@testing-library/dom";
+import "@testing-library/jest-dom";
+import {
   encodeHTML,
-  renderError,
-  flexLayout,
   getCardColors,
+  kFormatter,
+  parseBoolean,
+  renderError,
   wrapTextMultiline,
-} = require("../src/common/utils");
-
-const { queryByTestId } = require("@testing-library/dom");
+} from "../src/common/utils.js";
 
 describe("Test utils.js", () => {
   it("should test kFormatter", () => {
@@ -19,6 +18,23 @@ describe("Test utils.js", () => {
     expect(kFormatter(10000)).toBe("10k");
     expect(kFormatter(12345)).toBe("12.3k");
     expect(kFormatter(9900000)).toBe("9900k");
+  });
+
+  it("should test parseBoolean", () => {
+    expect(parseBoolean(true)).toBe(true);
+    expect(parseBoolean(false)).toBe(false);
+
+    expect(parseBoolean("true")).toBe(true);
+    expect(parseBoolean("false")).toBe(false);
+    expect(parseBoolean("True")).toBe(true);
+    expect(parseBoolean("False")).toBe(false);
+    expect(parseBoolean("TRUE")).toBe(true);
+    expect(parseBoolean("FALSE")).toBe(false);
+
+    expect(parseBoolean("1")).toBe(undefined);
+    expect(parseBoolean("0")).toBe(undefined);
+    expect(parseBoolean("")).toBe(undefined);
+    expect(parseBoolean(undefined)).toBe(undefined);
   });
 
   it("should test encodeHTML", () => {
@@ -32,7 +48,9 @@ describe("Test utils.js", () => {
     expect(
       queryByTestId(document.body, "message").children[0],
     ).toHaveTextContent(/Something went wrong/gim);
-    expect(queryByTestId(document.body, "message").children[1]).toBeEmpty(2);
+    expect(
+      queryByTestId(document.body, "message").children[1],
+    ).toBeEmptyDOMElement(2);
 
     // Secondary message
     document.body.innerHTML = renderError(
@@ -48,6 +66,7 @@ describe("Test utils.js", () => {
     let colors = getCardColors({
       title_color: "f00",
       text_color: "0f0",
+      ring_color: "0000ff",
       icon_color: "00f",
       bg_color: "fff",
       border_color: "fff",
@@ -57,6 +76,7 @@ describe("Test utils.js", () => {
       titleColor: "#f00",
       textColor: "#0f0",
       iconColor: "#00f",
+      ringColor: "#0000ff",
       bgColor: "#fff",
       borderColor: "#fff",
     });
@@ -75,6 +95,7 @@ describe("Test utils.js", () => {
       titleColor: "#2f80ed",
       textColor: "#0f0",
       iconColor: "#00f",
+      ringColor: "#2f80ed",
       bgColor: "#fff",
       borderColor: "#e4e2e2",
     });
@@ -87,9 +108,29 @@ describe("Test utils.js", () => {
     expect(colors).toStrictEqual({
       titleColor: "#fff",
       textColor: "#9f9f9f",
+      ringColor: "#fff",
       iconColor: "#79ff97",
       bgColor: "#151515",
       borderColor: "#e4e2e2",
+    });
+  });
+
+  it("getCardColors: should return ring color equal to title color if not ring color is defined", () => {
+    let colors = getCardColors({
+      title_color: "f00",
+      text_color: "0f0",
+      icon_color: "00f",
+      bg_color: "fff",
+      border_color: "fff",
+      theme: "dark",
+    });
+    expect(colors).toStrictEqual({
+      titleColor: "#f00",
+      textColor: "#0f0",
+      iconColor: "#00f",
+      ringColor: "#f00",
+      bgColor: "#fff",
+      borderColor: "#fff",
     });
   });
 });
